+++ docs/source/conf.py
@@ -102,10 +102,6 @@ html_context = {
     "doc_path": "/docs/source/",
 }
 
-rst_epilog = """
-.. |ProjectVersion| replace::{versionnum}
-""".format(versionnum=release)
-
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".