+++ scripts/para.py
@@ -1,0 +1,9 @@
+# coding=utf-8
+"""
+Script entry for the CLI executable 'para' for pyinstaller
+"""
+
+if __name__ == '__main__':
+    import paralang_cli
+
+    paralang_cli.scripts.paraproj.cli_run()